move era_validator_weights_for_block into linear chain module
